Split calculation and model
